style: Routine타입 파일 분할.
