+++ CodeStarterCamp_Week4/main.swift
@@ -38,53 +38,6 @@ struct BodyCondition {
     }
 }
 
-struct Routine {
-    let name: String
-    var exercises = [Exercise]()
-    
-    init(name: String) {
-        self.name = name
-    }
-    init(name: String, exercises: Exercise...) {
-        self.name = name
-        self.exercises = exercises
-    }
-    
-    mutating func appendExercise(_ exercise: Exercise) {
-        exercises.append(exercise)
-    }
-    
-    mutating func insertExercise(_ exercise: Exercise, at point: Int) {
-        if point < exercises.count {
-            exercises.insert(exercise, at: point)
-        }
-    }
-    
-    mutating func removeExercise(at point: Int) {
-        if point < exercises.count {
-            exercises.remove(at: point)
-        }
-    }
-    
-    func startRoutine(based bodyCondition: BodyCondition) -> BodyCondition {
-        var exerciserBodyCondition = bodyCondition
-        print("----------------------------------")
-        if exercises.isEmpty {
-            print("Don't start \(name) Routine because \(name) Routine is empty")
-            return exerciserBodyCondition
-        } else {
-            print("\(name) Routine Start")
-        }
-        print("----------------------------------")
-        for exercise in exercises {
-            print("\(exercise.name)")
-            exerciserBodyCondition = exercise.action(exerciserBodyCondition)
-        }
-        exerciserBodyCondition.informBodyCondition()
-        return exerciserBodyCondition
-    }
-}
-
 struct Exercise {
     let name: String
     let action: (BodyCondition) -> BodyCondition
